requirements.txt
discord.py==2.3.2
anthropic==0.31.0
httpx==0.24.1
GitPython==3.1.40
python-dotenv==1.0.0
requests==2.31.0
